@@ -1,5 +1,5 @@
 const AddIcon = props => {
-  return <button className={'badge badge-add'} onClick={props.onClick}><i className="glyphicon glyphicon-plus"></i> Add Word</button>
+  return <div className="add-button-container"><button className={'badge badge-add'} onClick={props.onClick}><i className="glyphicon glyphicon-plus"></i> Add Word</button></div>
 }
 
 export default AddIcon;
